+++ cmd/frontend/internal/frontend/v1/buf.gen.yaml
